requirements.txt
Django==4.0.1
django-ninja==0.13.2
pytest==6.2.5
pytest-django==4.5.2